pyproject.toml
[project]
name = "vault-search-replace"
version = "0.1.0"
description = "A simple search and replace command line utility for Hashicorp Vault"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "hvac>=2.3.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "ruff>=0.6.9",
    "typer>=0.12.5",
]

[project.scripts]
vault-search-replace = "vault_search_replace:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.ruff.lint]
select = []
